// src/model/vehicle.model.ts
import { Publication } from "./publication.model";


export interface Vehicle {
      vehicleId: string;
      publication?: Publication;
      image1?: string;
      image2?: string;
      image3?: string;
      image4?: string;
      brand: string;
      model: string;
      vehicleYear: string;
      doorQuantity: string;
      fuelType: string;
      kilometers: string;
      color: string;
      steering: string;
      transmission: string;
      tractionControl: string;
      armored: string;
      hasAbsBrakes: string;
      hasAirConditioning: string;
      hasAlarm: string;
      hasAlloyWheels: string;
      hasAmfmRadio: string;
      hasAuxiliaryPort: string;
      itemCondition: string;
      licensePlate: string;
      passengerCapacity: string;
      trim: string;
      singleOwner: string;
      hasWindscreenWiper: string;
      hasUsb: string;
      hasSteeringWheelControl: string;
      hasPowerDoorLocks: string;
      hasPowerWindows: string;
      hasRainSensor: string;
      hasHeightAdjustableDriverSeat: string;
      hasRearFoglights: string;
      hasSlidingRoof: string;
      hasLeatherUpholstery: string;
      hasLightOnReminder: string;
      hasMp3Player: string;
      hasOnboardComputer: string;
      hasParkingSensor: string;
      hasPassengerAirbag: string;
      vehicleBodyType: string;
      height: string;
      width: string;
      distanceBetweenAxes: string;
      fuelCapacity: string;
      length: string;
      valvesPerCylinder: string;
      gearNumber: string;
      hasCurtainAirbag: string;
      hasFogLight: string;
      hasRearFoldingSeat: string;
      hasElectricMirrors: string;
      hasAutopilot: string; 
}